--- docker-compose-dev.yml
@@ -1,9 +1,0 @@
-version: "3"
-services:
-  backend:
-    image: openslides-backend-dev
-    ports:
-      - "9002:9002"
-      - "9003:9003"
-    volumes:
-      - ./openslides_backend:/srv/code/openslides_backends
