+++ setup.cfg
@@ -21,3 +21,4 @@ exclude_lines =
     pragma: no cover
     raise NotImplementedError
     return NotImplemented
+fail_under = 84
