--- tests/Dockerfile
@@ -1,1 +1,0 @@
-
