+++ tests/unit/services/test_authentication_adapter.py
@@ -1,0 +1,7 @@
+from unittest import TestCase
+
+
+class AuthentificationTester(TestCase):
+    # TODO: write unit tests for auth adapter. Mock `requests.post` to simulate
+    # different results.
+    pass
