+++ tests/unit/services/test_database_adapter.py
