+++ tests/util.py
@@ -1,0 +1,22 @@
+from openslides_backend.shared.patterns import (
+    KEYSEPARATOR,
+    Collection,
+    FullQualifiedField,
+    FullQualifiedId,
+)
+
+
+def get_fqid(value: str) -> FullQualifiedId:
+    """
+    Returns a FullQualifiedId parsed from the given value.
+    """
+    collection, id = value.split(KEYSEPARATOR)
+    return FullQualifiedId(Collection(collection), int(id))
+
+
+def get_fqfield(value: str) -> FullQualifiedField:
+    """
+    Returns a FullQualifiedField parsed from the given value.
+    """
+    collection, id, field = value.split(KEYSEPARATOR)
+    return FullQualifiedField(Collection(collection), int(id), field)
